Refactor: moved makeMovieWithSettings function to Player
