+++ cc3d/player5/.gitignore
@@ -1,0 +1,2 @@
+.idea/*
+**/__pycache__
